requirements.txt
pillow==9.3.0
loguru==0.6.0
